+++ src/hardware/ouros/reducer.rs
@@ -29,7 +29,7 @@ impl HwInput for ReducerInput {}
 
 #[derive(Default)]
 pub struct ReducerStat {
-    pub busy_cycles: u8,
+    pub busy_cycles: u32,
     pub holder_contents: Vec<Option<App>>,
 }
